step 1: pull out task creators
