+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/RecordCollectorImpl.java
@@ -262,7 +262,7 @@ public class RecordCollectorImpl implements RecordCollector {
         if (eosEnabled) {
             streamsProducer.abortTransaction();
         }
-        streamsProducer.close();
+        streamsProducer.flush();
 
         checkForException();
     }
