+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/StreamThread.java
@@ -23,7 +23,6 @@ import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.InvalidOffsetException;
 import org.apache.kafka.clients.producer.Producer;
-import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.Metric;
 import org.apache.kafka.common.MetricName;
@@ -49,11 +48,8 @@ import org.apache.kafka.streams.state.internals.ThreadCache;
 import org.slf4j.Logger;
 
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -61,8 +57,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
-
-import static org.apache.kafka.streams.StreamsConfig.EXACTLY_ONCE;
 
 public class StreamThread extends Thread {
 
@@ -247,235 +241,6 @@ public class StreamThread extends Thread {
         return assignmentErrorCode.get();
     }
 
-    static abstract class AbstractTaskCreator<T extends Task> {
-        final String applicationId;
-        final InternalTopologyBuilder builder;
-        final StreamsConfig config;
-        final StreamsMetricsImpl streamsMetrics;
-        final StateDirectory stateDirectory;
-        final ChangelogReader storeChangelogReader;
-        final Time time;
-        final Logger log;
-
-        AbstractTaskCreator(final InternalTopologyBuilder builder,
-                            final StreamsConfig config,
-                            final StreamsMetricsImpl streamsMetrics,
-                            final StateDirectory stateDirectory,
-                            final ChangelogReader storeChangelogReader,
-                            final Time time,
-                            final Logger log) {
-            this.applicationId = config.getString(StreamsConfig.APPLICATION_ID_CONFIG);
-            this.builder = builder;
-            this.config = config;
-            this.streamsMetrics = streamsMetrics;
-            this.stateDirectory = stateDirectory;
-            this.storeChangelogReader = storeChangelogReader;
-            this.time = time;
-            this.log = log;
-        }
-
-        public InternalTopologyBuilder builder() {
-            return builder;
-        }
-
-        public StateDirectory stateDirectory() {
-            return stateDirectory;
-        }
-
-        Collection<T> createTasks(final Consumer<byte[], byte[]> consumer,
-                                  final Map<TaskId, Set<TopicPartition>> tasksToBeCreated) {
-            final List<T> createdTasks = new ArrayList<>();
-            for (final Map.Entry<TaskId, Set<TopicPartition>> newTaskAndPartitions : tasksToBeCreated.entrySet()) {
-                final TaskId taskId = newTaskAndPartitions.getKey();
-                final Set<TopicPartition> partitions = newTaskAndPartitions.getValue();
-                final T task = createTask(consumer, taskId, partitions);
-                if (task != null) {
-                    log.trace("Created task {} with assigned partitions {}", taskId, partitions);
-                    createdTasks.add(task);
-                }
-
-            }
-            return createdTasks;
-        }
-
-        abstract T createTask(final Consumer<byte[], byte[]> consumer, final TaskId id, final Set<TopicPartition> partitions);
-
-        void close() {}
-    }
-
-    static class TaskCreator extends AbstractTaskCreator<StreamTask> {
-        private final String threadId;
-        private final ThreadCache cache;
-        private final Producer<byte[], byte[]> threadProducer;
-        private final KafkaClientSupplier clientSupplier;
-        final Map<TaskId, Producer<byte[], byte[]>> taskProducers;
-        private final Sensor createTaskSensor;
-
-        TaskCreator(final InternalTopologyBuilder builder,
-                    final StreamsConfig config,
-                    final StreamsMetricsImpl streamsMetrics,
-                    final StateDirectory stateDirectory,
-                    final ChangelogReader storeChangelogReader,
-                    final ThreadCache cache,
-                    final Time time,
-                    final KafkaClientSupplier clientSupplier,
-                    final Map<TaskId, Producer<byte[], byte[]>> taskProducers,
-                    final String threadId,
-                    final Logger log) {
-            super(
-                builder,
-                config,
-                streamsMetrics,
-                stateDirectory,
-                storeChangelogReader,
-                time,
-                log);
-
-            final boolean eosEnabled = EXACTLY_ONCE.equals(config.getString(StreamsConfig.PROCESSING_GUARANTEE_CONFIG));
-            if (!eosEnabled) {
-                final Map<String, Object> producerConfigs = config.getProducerConfigs(getThreadProducerClientId(threadId));
-                log.info("Creating thread producer client");
-                this.threadProducer = clientSupplier.getProducer(producerConfigs);
-            } else {
-                this.threadProducer = null;
-            }
-            this.taskProducers = taskProducers;
-
-            this.cache = cache;
-            this.threadId = threadId;
-            this.clientSupplier = clientSupplier;
-
-            this.createTaskSensor = ThreadMetrics.createTaskSensor(threadId, streamsMetrics);
-        }
-
-        @Override
-        StreamTask createTask(final Consumer<byte[], byte[]> mainConsumer,
-                              final TaskId taskId,
-                              final Set<TopicPartition> partitions) {
-            createTaskSensor.record();
-
-            final String threadIdPrefix = String.format("stream-thread [%s] ", Thread.currentThread().getName());
-            final String logPrefix = threadIdPrefix + String.format("%s [%s] ", "task", taskId);
-            final LogContext logContext = new LogContext(logPrefix);
-
-            final ProcessorTopology topology = builder.buildSubtopology(taskId.topicGroupId);
-
-            final ProcessorStateManager stateManager = new ProcessorStateManager(
-                taskId,
-                partitions,
-                Task.TaskType.ACTIVE,
-                stateDirectory,
-                topology.storeToChangelogTopic(),
-                storeChangelogReader,
-                logContext);
-
-            if (threadProducer == null) {
-                // create one producer per task for EOS
-                // TODO: after KIP-447 this would be removed
-                final Map<String, Object> producerConfigs = config.getProducerConfigs(getTaskProducerClientId(threadId, taskId));
-                producerConfigs.put(ProducerConfig.TRANSACTIONAL_ID_CONFIG, applicationId + "-" + taskId);
-                log.info("Creating producer client for task {}", taskId);
-                taskProducers.put(taskId, clientSupplier.getProducer(producerConfigs));
-            }
-            final RecordCollector recordCollector = new RecordCollectorImpl(
-                logContext,
-                taskId,
-                mainConsumer,
-                threadProducer != null ?
-                    new StreamsProducer(logContext, threadProducer) :
-                    new StreamsProducer(logContext, taskProducers.get(taskId), applicationId, taskId),
-                config.defaultProductionExceptionHandler(),
-                EXACTLY_ONCE.equals(config.getString(StreamsConfig.PROCESSING_GUARANTEE_CONFIG)),
-                streamsMetrics);
-
-            return new StreamTask(
-                taskId,
-                partitions,
-                topology,
-                mainConsumer,
-                config,
-                streamsMetrics,
-                stateDirectory,
-                cache,
-                time,
-                stateManager,
-                recordCollector);
-        }
-
-        public void close() {
-            if (threadProducer != null) {
-                try {
-                    threadProducer.close();
-                } catch (final Throwable e) {
-                    log.error("Failed to close producer due to the following error:", e);
-                }
-            }
-        }
-    }
-
-    static class StandbyTaskCreator extends AbstractTaskCreator<StandbyTask> {
-        private final Sensor createTaskSensor;
-
-        StandbyTaskCreator(final InternalTopologyBuilder builder,
-                           final StreamsConfig config,
-                           final StreamsMetricsImpl streamsMetrics,
-                           final StateDirectory stateDirectory,
-                           final ChangelogReader storeChangelogReader,
-                           final Time time,
-                           final String threadId,
-                           final Logger log) {
-            super(
-                builder,
-                config,
-                streamsMetrics,
-                stateDirectory,
-                storeChangelogReader,
-                time,
-                log);
-            createTaskSensor = ThreadMetrics.createTaskSensor(threadId, streamsMetrics);
-        }
-
-        @Override
-        StandbyTask createTask(final Consumer<byte[], byte[]> consumer,
-                               final TaskId taskId,
-                               final Set<TopicPartition> partitions) {
-            createTaskSensor.record();
-
-            final String threadIdPrefix = String.format("stream-thread [%s] ", Thread.currentThread().getName());
-            final String logPrefix = threadIdPrefix + String.format("%s [%s] ", "standby-task", taskId);
-            final LogContext logContext = new LogContext(logPrefix);
-
-            final ProcessorTopology topology = builder.buildSubtopology(taskId.topicGroupId);
-
-            if (topology.hasStateWithChangelogs()) {
-                final ProcessorStateManager stateManager = new ProcessorStateManager(
-                    taskId,
-                    partitions,
-                    Task.TaskType.STANDBY,
-                    stateDirectory,
-                    topology.storeToChangelogTopic(),
-                    storeChangelogReader,
-                    logContext);
-
-                return new StandbyTask(
-                    taskId,
-                    partitions,
-                    topology,
-                    config,
-                    streamsMetrics,
-                    stateManager,
-                    stateDirectory);
-            } else {
-                log.trace(
-                    "Skipped standby task {} with assigned partitions {} " +
-                        "since it does not have any state stores to materialize",
-                    taskId, partitions
-                );
-                return null;
-            }
-        }
-    }
-
     private final Time time;
     private final Logger log;
     private final String logPrefix;
@@ -544,11 +309,7 @@ public class StreamThread extends Thread {
 
         final ThreadCache cache = new ThreadCache(logContext, cacheSizeBytes, streamsMetrics);
 
-        final Map<TaskId, Producer<byte[], byte[]>> taskProducers = new HashMap<>();
-
-        // TODO: refactor `TaskCreator` into `TaskManager`;
-        //  this will allow to reduce the surface area of `taskProducers` that is passed to many classes atm
-        final TaskCreator activeTaskCreator = new TaskCreator(
+        final ActiveTaskCreator activeTaskCreator = new ActiveTaskCreator(
             builder,
             config,
             streamsMetrics,
@@ -557,7 +318,6 @@ public class StreamThread extends Thread {
             cache,
             time,
             clientSupplier,
-            taskProducers,
             threadId,
             log);
         final StandbyTaskCreator standbyTaskCreator = new StandbyTaskCreator(
@@ -576,7 +336,6 @@ public class StreamThread extends Thread {
             streamsMetrics,
             activeTaskCreator,
             standbyTaskCreator,
-            taskProducers,
             builder,
             adminClient
         );
@@ -602,8 +361,8 @@ public class StreamThread extends Thread {
         final StreamThread streamThread = new StreamThread(
             time,
             config,
-            activeTaskCreator.threadProducer,
-            taskProducers,
+            activeTaskCreator.threadProducer(),
+            activeTaskCreator.taskProducers(),
             adminClient,
             mainConsumer,
             restoreConsumer,
@@ -686,11 +445,11 @@ public class StreamThread extends Thread {
         }
     }
 
-    private static String getTaskProducerClientId(final String threadClientId, final TaskId taskId) {
+    static String getTaskProducerClientId(final String threadClientId, final TaskId taskId) {
         return threadClientId + "-" + taskId + "-producer";
     }
 
-    private static String getThreadProducerClientId(final String threadClientId) {
+    static String getThreadProducerClientId(final String threadClientId) {
         return threadClientId + "-producer";
     }
